skin_cancer_classifier/requirements.txt
torch
torchvision
albumentations
scikit-learn
matplotlib
seaborn
kagglehub
opencv-python-headless
pandas
tqdm
Pillow
numpy
